[refactor] header date 분리
