+++ FE/src/components/Common/MenuBar/HeaderButtons/header-buttons.scss
@@ -1,0 +1,37 @@
+@use '../../../../styles/values' as *;
+@use '../../../../styles/mixins' as *;
+@use '../../../../styles/mediaQuery' as *;
+
+@mixin navigationButton($type) {
+  @include headerNavigationButtons($type);
+
+  @include mobile {
+    .navigation__content {
+      display: none;
+    }
+    .fas.fa-arrow-right {
+      right: 5%;
+    }
+  }
+}
+
+.menubar__buttons {
+  @include headerButtons;
+  .back__navBtn {
+    @include navigationButton(left);
+  }
+  .setting__navBtn {
+    @include navigationButton(right);
+  }
+  @include mobile {
+    .back__navBtn,
+    .setting__navBtn {
+      font-size: 1rem;
+    }
+  }
+}
+@include mobile {
+  .menubar__buttons {
+    gap: 2em;
+  }
+}
