+++ FE/src/components/Common/MenuBar/index.tsx
@@ -1,41 +1,15 @@
-import React, { useRef, useEffect, useCallback } from 'react';
+import React, { useRef, useCallback } from 'react';
 
 import { useHistory } from 'react-router-dom';
-import { useDateInfoData } from '../../../store/DateInfo/dateInfoHook';
-import HeaderButton from '../HeaderButton';
+
+import HeaderButtons from './HeaderButtons';
+import HeaderDate from './HeaderDate';
 import './menubar.scss';
-import CalculateDate from '../../../util/calculateDate';
 
 const MenuBar = ({ id, setModal, pageType }) => {
   const history = useHistory();
-  const DateInfo = useDateInfoData(store => store.nowCalendarInfo);
-  const setDateInfo = useDateInfoData(store => store.setCalendarInfo);
 
   const allBtnRef = useRef();
-  const btnNextRef = useRef();
-  const btnPrevRef = useRef();
-  const calMonRef = useRef();
-  const calYearRef = useRef();
-
-  const setYearMonth = useCallback((year, month) => {
-    const yy = year;
-    const mm = month;
-
-    calMonRef.current.textContent = CalculateDate.monList[mm];
-    calYearRef.current.textContent = yy;
-  }, []);
-
-  const onClickNextMonth = useCallback(() => {
-    const info = CalculateDate.nextMonth();
-    setYearMonth(info.getFullYear(), info.getMonth());
-    setDateInfo(info.getFullYear(), info.getMonth());
-  }, []);
-
-  const onClickPrevMonth = useCallback(() => {
-    const info = CalculateDate.prevMonth();
-    setYearMonth(info.getFullYear(), info.getMonth());
-    setDateInfo(info.getFullYear(), info.getMonth());
-  }, []);
 
   const onClickIcon = useCallback(event => {
     history.push({
@@ -45,14 +19,6 @@ const MenuBar = ({ id, setModal, pageType }) => {
       },
     });
   }, []);
-
-  const onClickPayment = useCallback(() => {
-    setModal(true);
-  }, []);
-
-  useEffect(() => {
-    setYearMonth(DateInfo.year, DateInfo.month);
-  }, [pageType]);
 
   return (
     <header className="menubar__header">
@@ -67,52 +33,7 @@ const MenuBar = ({ id, setModal, pageType }) => {
             List
           </span>
         </div>
-        {pageType === 'transaction' ? (
-          <HeaderButton
-            buttonType="transaction"
-            isChecked
-            onClickIcon={onClickIcon}
-          />
-        ) : (
-          <HeaderButton
-            buttonType="transaction"
-            isChecked={false}
-            onClickIcon={onClickIcon}
-          />
-        )}
-        {pageType === 'calendar' ? (
-          <HeaderButton
-            buttonType="calendar"
-            isChecked
-            onClickIcon={onClickIcon}
-          />
-        ) : (
-          <HeaderButton
-            buttonType="calendar"
-            isChecked={false}
-            onClickIcon={onClickIcon}
-          />
-        )}
-        {pageType === 'chart' ? (
-          <HeaderButton
-            buttonType="chart"
-            isChecked
-            onClickIcon={onClickIcon}
-          />
-        ) : (
-          <HeaderButton
-            buttonType="chart"
-            isChecked={false}
-            onClickIcon={onClickIcon}
-          />
-        )}
-
-        <HeaderButton
-          buttonType="paymentMethod"
-          isChecked={false}
-          onClickIcon={onClickPayment}
-        />
-
+        <HeaderButtons id={id} pageType={pageType} setModal={setModal} />
         <div
           className={
             pageType === '/setting'
@@ -129,27 +50,7 @@ const MenuBar = ({ id, setModal, pageType }) => {
         </div>
       </div>
 
-      <div className="menubar__ctrBox">
-        <button
-          type="button"
-          title="prev"
-          className="btn-cal prev"
-          onClick={onClickPrevMonth}
-          ref={btnPrevRef}
-        />
-        <div className="year-month">
-          <span className="cal-year" ref={calYearRef} />
-          <span className="cal-month" ref={calMonRef} />
-        </div>
-
-        <button
-          type="button"
-          title="next"
-          className="btn-cal next"
-          onClick={onClickNextMonth}
-          ref={btnNextRef}
-        />
-      </div>
+      <HeaderDate pageType={pageType} />
     </header>
   );
 };
